+++ src/tabViews/charts/SectionBy.tsx
@@ -2,7 +2,8 @@ import Autocomplete from "@mui/material/Autocomplete";
 import Box from "@mui/material/Box";
 import TextField from "@mui/material/TextField";
 import { HTMLAttributes, SyntheticEvent, useMemo, useState } from "react";
-import { useAllTabs, useFilterDispatch } from "src/contexts/dataSelectors";
+import { useFilterDispatch } from "src/contexts/DataProvider";
+import { useAllTabs } from "src/contexts/selectors";
 import { TAB_PROPERTIES } from "src/utils/chrome";
 import { FILTER_TAB_PROPERTIES } from "src/utils/filterTabs";
 import TabPropertyIcon, { getTabPropertyLabel } from "../window/TabPropertyIcon";
